requirements.txt
opencv-python
langchain
langchain-openai
langchain_google_genai
langchain-community
python-dotenv
pyaudio
soundfile
SpeechRecognition
python-dotenv
git+https://github.com/openai/whisper.git
